Finish clock component
